Stop validating disabled questions
