+++ datacapture/src/main/java/com/google/android/fhir/datacapture/enablement/EnablementEvaluator.kt
@@ -24,10 +24,12 @@ import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
 
 /**
- * Evaluator for the enablement status of a [Questionnaire.QuestionnaireItemComponent]. Uses the
- * `enableWhen` constraints and the `enableBehavior` value defined in the
- * [Questionnaire.QuestionnaireItemComponent]. Also depends on the answers (or lack thereof)
- * captured in the specified [QuestionnaireResponse.QuestionnaireResponseItemComponent] s.
+ * Evaluator for the enablement status of a [Questionnaire.QuestionnaireItemComponent].
+ *
+ * This is done by locating the relevant [QuestionnaireResponse.QuestionnaireResponseItemComponent]s
+ * specified by the linkIds in the `enableWhen` constraints, and checking if the answers (or lack
+ * thereof) satisfy the criteria in the `enableWhen` constraints. The `enableBehavior` value is then
+ * used to combine the evaluation results of different `enableWhen` constraints.
  *
  * For example, the following `enableWhen` constraint in a
  * [Questionnaire.QuestionnaireItemComponent]
@@ -41,11 +43,14 @@ import org.hl7.fhir.r4.model.QuestionnaireResponse
  *     ],
  * ```
  * specifies that the [Questionnaire.QuestionnaireItemComponent] should be enabled only if the
- * question with ID `vitaminKgiven` has been answered.
+ * question with linkId `vitaminKgiven` has been answered.
  *
  * The enablement status typically determines whether the [Questionnaire.QuestionnaireItemComponent]
  * is shown or hidden. However, it is also possible that only user interaction is enabled or
  * disabled (e.g. grayed out) with the [Questionnaire.QuestionnaireItemComponent] always shown.
+ *
+ * The evaluator does not track the changes in the `questionnaire` and `questionnaireResponse`.
+ * Therefore, a new evaluator should be created if they were modified.
  *
  * For more information see
  * [Questionnaire.item.enableWhen](https://www.hl7.org/fhir/questionnaire-definitions.html#Questionnaire.item.enableWhen)
@@ -53,7 +58,60 @@ import org.hl7.fhir.r4.model.QuestionnaireResponse
  * [Questionnaire.item.enableBehavior](https://www.hl7.org/fhir/questionnaire-definitions.html#Questionnaire.item.enableBehavior)
  * .
  */
-internal object EnablementEvaluator {
+internal class EnablementEvaluator(
+  val questionnaire: Questionnaire,
+  val questionnaireResponse: QuestionnaireResponse
+) {
+  /**
+   * The pre-order traversal trace of the items in the [QuestionnaireResponse]. This essentially
+   * represents the order in which all items are displayed in the UI.
+   */
+  private val questionnaireResponseItemPreOrderList =
+    mutableListOf<QuestionnaireResponse.QuestionnaireResponseItemComponent>()
+
+  init {
+    /**
+     * Adds all items in the [QuestionnaireResponse] to the pre-order list. Note that each
+     * questionnaire response item may either have child items (in the case of a group type
+     * question) or have answer items with nested questions.
+     */
+    fun buildPreOrderList(item: QuestionnaireResponse.QuestionnaireResponseItemComponent) {
+      questionnaireResponseItemPreOrderList.add(item)
+      for (child in item.item) {
+        buildPreOrderList(child)
+      }
+      for (answer in item.answer) {
+        for (answerItem in answer.item) {
+          buildPreOrderList(answerItem)
+        }
+      }
+    }
+
+    for (item in questionnaireResponse.item) {
+      buildPreOrderList(item)
+    }
+  }
+
+  /** The map from each item in the [QuestionnaireResponse] to its parent. */
+  private val questionnaireResponseItemParentMap =
+    mutableMapOf<
+      QuestionnaireResponse.QuestionnaireResponseItemComponent,
+      QuestionnaireResponse.QuestionnaireResponseItemComponent
+    >()
+
+  init {
+    /** Adds each child-parent pair in the [QuestionnaireResponse] to the parent map. */
+    fun buildParentList(item: QuestionnaireResponse.QuestionnaireResponseItemComponent) {
+      for (child in item.item) {
+        questionnaireResponseItemParentMap[child] = item
+        buildParentList(child)
+      }
+    }
+
+    for (item in questionnaireResponse.item) {
+      buildParentList(item)
+    }
+  }
 
   /**
    * Returns whether [questionnaireItem] should be enabled.
@@ -66,12 +124,6 @@ internal object EnablementEvaluator {
   fun evaluate(
     questionnaireItem: Questionnaire.QuestionnaireItemComponent,
     questionnaireResponseItem: QuestionnaireResponse.QuestionnaireResponseItemComponent,
-    questionnaireResponse: QuestionnaireResponse,
-    questionnaireResponseItemRetriever:
-      (
-        item: QuestionnaireResponse.QuestionnaireResponseItemComponent,
-        linkId: String
-      ) -> QuestionnaireResponse.QuestionnaireResponseItemComponent?
   ): Boolean {
     val enableWhenList = questionnaireItem.enableWhen
     val enableWhenExpression = questionnaireItem.enableWhenExpression
@@ -92,7 +144,6 @@ internal object EnablementEvaluator {
       return evaluateEnableWhen(
         enableWhenList.single(),
         questionnaireResponseItem,
-        questionnaireResponseItemRetriever
       )
     }
 
@@ -102,43 +153,79 @@ internal object EnablementEvaluator {
     // enabled if ANY `enableWhen` constraint is satisfied.
     return when (val value = questionnaireItem.enableBehavior) {
       Questionnaire.EnableWhenBehavior.ALL ->
-        enableWhenList.all {
-          evaluateEnableWhen(it, questionnaireResponseItem, questionnaireResponseItemRetriever)
-        }
+        enableWhenList.all { evaluateEnableWhen(it, questionnaireResponseItem) }
       Questionnaire.EnableWhenBehavior.ANY ->
-        enableWhenList.any {
-          evaluateEnableWhen(it, questionnaireResponseItem, questionnaireResponseItemRetriever)
-        }
+        enableWhenList.any { evaluateEnableWhen(it, questionnaireResponseItem) }
       else -> throw IllegalStateException("Unrecognized enable when behavior $value")
     }
   }
-}
 
-/**
- * Returns whether the `enableWhen` constraint is satisfied.
- *
- * @param questionnaireResponseItemRetriever function that returns the [QuestionnaireResponse.Item]
- * with the `linkId`, or null if there isn't one.
- */
-private fun evaluateEnableWhen(
-  enableWhen: Questionnaire.QuestionnaireItemEnableWhenComponent,
-  questionnaireResponseItem: QuestionnaireResponse.QuestionnaireResponseItemComponent,
-  questionnaireResponseItemRetriever:
-    (
-      item: QuestionnaireResponse.QuestionnaireResponseItemComponent,
-      linkId: String
-    ) -> QuestionnaireResponse.QuestionnaireResponseItemComponent?
-): Boolean {
-  val questionnaireResponseItem =
-    questionnaireResponseItemRetriever(questionnaireResponseItem, enableWhen.question)
-  return if (Questionnaire.QuestionnaireItemOperator.EXISTS == enableWhen.operator) {
-    (questionnaireResponseItem == null || questionnaireResponseItem.answer.isEmpty()) !=
-      enableWhen.answerBooleanType.booleanValue()
-  } else {
-    // The `enableWhen` constraint evaluates to true if at least one answer has a value that
-    // satisfies the `enableWhen` operator and answer, with the exception of the `Exists` operator.
-    // See https://www.hl7.org/fhir/valueset-questionnaire-enable-operator.html.
-    questionnaireResponseItem?.answer?.any { enableWhen.predicate(it) } ?: false
+  /**
+   * Returns whether the `enableWhen` constraint is satisfied.
+   *
+   * @param questionnaireResponseItemRetriever function that returns the
+   * [QuestionnaireResponse.Item] with the `linkId`, or null if there isn't one.
+   */
+  private fun evaluateEnableWhen(
+    enableWhen: Questionnaire.QuestionnaireItemEnableWhenComponent,
+    questionnaireResponseItem: QuestionnaireResponse.QuestionnaireResponseItemComponent,
+  ): Boolean {
+    val target =
+      findEnableWhenQuestionnaireResponseItem(questionnaireResponseItem, enableWhen.question)
+    return if (Questionnaire.QuestionnaireItemOperator.EXISTS == enableWhen.operator) {
+      (target == null || target.answer.isEmpty()) != enableWhen.answerBooleanType.booleanValue()
+    } else {
+      // The `enableWhen` constraint evaluates to true if at least one answer has a value that
+      // satisfies the `enableWhen` operator and answer, with the exception of the `Exists`
+      // operator.
+      // See https://www.hl7.org/fhir/valueset-questionnaire-enable-operator.html.
+      target?.answer?.any { enableWhen.predicate(it) } ?: false
+    }
+  }
+
+  /**
+   * Find a questionnaire response item in [QuestionnaireResponse] with the given `linkId` starting
+   * from the `origin`.
+   *
+   * This is used by the enableWhen logic to evaluate if a question should be enabled/displayed.
+   *
+   * If multiple questionnaire response items are present for the same question (same linkId),
+   * either as a result of repeated group or nested question under repeated answers, this returns
+   * the nearest question occurrence reachable by tracing first the "ancestor" axis and then the
+   * "preceding" axis and then the "following" axis.
+   *
+   * See
+   * https://www.hl7.org/fhir/questionnaire-definitions.html#Questionnaire.item.enableWhen.question.
+   */
+  private fun findEnableWhenQuestionnaireResponseItem(
+    origin: QuestionnaireResponse.QuestionnaireResponseItemComponent,
+    linkId: String,
+  ): QuestionnaireResponse.QuestionnaireResponseItemComponent? {
+    // Find the nearest ancestor with the linkId
+    var parent = questionnaireResponseItemParentMap[origin]
+    while (parent != null) {
+      if (parent.linkId == linkId) {
+        return parent
+      }
+      parent = questionnaireResponseItemParentMap[parent]
+    }
+
+    // Find the nearest item preceding the origin
+    val itemIndex = questionnaireResponseItemPreOrderList.indexOf(origin)
+    for (index in itemIndex - 1 downTo 0) {
+      if (questionnaireResponseItemPreOrderList[index].linkId == linkId) {
+        return questionnaireResponseItemPreOrderList[index]
+      }
+    }
+
+    // Find the nearest item succeeding the origin
+    for (index in itemIndex + 1 until questionnaireResponseItemPreOrderList.size) {
+      if (questionnaireResponseItemPreOrderList[index].linkId == linkId) {
+        return questionnaireResponseItemPreOrderList[index]
+      }
+    }
+
+    return null
   }
 }
